requirements.txt
# Core UI/agent framework
chainlit

# HTTP + utilities
requests
pyyaml

# Wikipedia API wrapper
wikipedia

# LLM SDKs
openai>=1.0.0
pydantic>=2.0.0

# LlamaIndex core and integrations
llama-index
llama-index-embeddings-huggingface
llama-index-llms-openai

# Embedding model backend
sentence-transformers
torch
